+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * [#151](https://github.com/exercism/cli/pull/151): Expand '~' in config path to home directory - [@lcowell](https://github.com/lcowell)
 * Stop supporting legacy config files (~/.exercism.go)
 * Deleted deprecated login/logout commands
+* Deleted deprecated key names in config
 * Your contribution here
 
 ## v1.9.2 (Jan 11, 2015)